requirements.txt
click==8.1.7
pika==1.3.2
pillow==10.4.0
protobuf==5.27.2
psycopg2-binary==2.9.9
redis==5.0.6
setuptools==70.0.0
wheel==0.43.0
